Deduplicate project resource request types
